Introduce ToolCardPresenter interface to replace Mockposable mocks with FakeToolCardPresenter
