+++ app/src/testDebug/kotlin/org/cru/godtools/ui/dashboard/home/HomePresenterTest.kt
@@ -5,7 +5,6 @@ import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.firebase.remoteconfig.FirebaseRemoteConfig
-import com.jeppeman.mockposable.mockk.everyComposable
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.test.FakeNavigator
 import com.slack.circuit.test.test
@@ -38,8 +37,8 @@ import org.cru.godtools.ui.banner.FakeBannerPresenter
 import org.cru.godtools.ui.banner.tutorial.TutorialFeaturesBannerPresenter
 import org.cru.godtools.ui.dashboard.home.HomePresenter.UiEvent
 import org.cru.godtools.ui.tooldetails.ToolDetailsScreen
+import org.cru.godtools.ui.tools.FakeToolCardPresenter
 import org.cru.godtools.ui.tools.ToolCard
-import org.cru.godtools.ui.tools.ToolCardPresenter
 import org.cru.godtools.util.createToolIntent
 import org.greenrobot.eventbus.EventBus
 import org.junit.runner.RunWith
@@ -60,15 +59,7 @@ class HomePresenterTest {
         every { getLessonsFlow() } returns lessonsFlow
         every { getFavoriteToolsFlow() } returns toolsFlow
     }
-    private val toolCardPresenter: ToolCardPresenter = mockk {
-        everyComposable { present(tool = any(), eventSink = any()) }.answers {
-            ToolCard.UiState(
-                toolCode = firstArg<Tool>().code,
-                translation = randomTranslation(languageCode = Locale.ENGLISH),
-                eventSink = arg(5)
-            )
-        }
-    }
+    private val toolCardPresenter = FakeToolCardPresenter()
     private val tutorialFeaturesBannerPresenter = FakeBannerPresenter<TutorialFeaturesBannerPresenter.UiState>()
 
     private val navigator = FakeNavigator(HomeScreen)
@@ -182,8 +173,8 @@ class HomePresenterTest {
     fun `State - spotlightLessons - Event - Click`() = runTest {
         val lesson = randomTool(type = Tool.Type.LESSON, isHidden = false, isSpotlight = true)
         val translation = randomTranslation(lesson.code, languageCode = Locale.FRENCH)
-        everyComposable { toolCardPresenter.present(tool = lesson, eventSink = any()) }.answers {
-            ToolCard.UiState(toolCode = lesson.code, translation = translation, eventSink = arg(5))
+        toolCardPresenter.onPresent = { t, _, es ->
+            ToolCard.UiState(toolCode = t.code, translation = translation, eventSink = es)
         }
         lessonsFlow.emit(listOf(lesson))
